% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expanded.R
\name{cal_gini}
\alias{cal_gini}
\title{calculate gini coefficiency for a sample}
\usage{
cal_gini(x, corr = FALSE, na.rm = TRUE)
}
\arguments{
\item{x}{a vector}
}
\value{
a value
}
\description{
calculate gini coefficiency for a sample
}
