% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{create_experimentqc_object}
\alias{create_experimentqc_object}
\title{Create a new experimentQC object}
\usage{
create_experimentqc_object(samqc_obj, coldata = NULL, refcond)
}
\arguments{
\item{samqc_obj}{a sampleQC object}

\item{coldata}{a data frame of coldata for DESeq2}

\item{refcond}{the reference condition, eg. D4}
}
\value{
An object of class sampleQC
}
\description{
Create a new experimentQC object
}
