% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{create_primaryqc_object}
\alias{create_primaryqc_object}
\title{Create a new primaryQC object}
\usage{
create_primaryqc_object(samples)
}
\arguments{
\item{samples}{a list of SGE objects}
}
\value{
An object of class primaryQC
}
\description{
Create a new primaryQC object
}
