% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_reports.R
\name{create_qc_reports}
\alias{create_qc_reports}
\title{create QC reports}
\usage{
create_qc_reports(
  samplesheet = NULL,
  samqc_cutoffs = NULL,
  qctype = "screen",
  qcdir = NULL
)
}
\arguments{
\item{samplesheet}{the path of sample sheet file}

\item{samqc_cutoffs}{the path of sample QC cutoffs (created by function qcout_cutoffs)}

\item{qctype}{screen or plasmid}

\item{qcdir}{the directory of QC plots and outs}
}
\description{
create QC reports
}
