% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{create_sampleqc_object}
\alias{create_sampleqc_object}
\title{Create a new sampleQC object}
\usage{
create_sampleqc_object(samples)
}
\arguments{
\item{samples}{a list of SGE objects}
}
\value{
An object of class sampleQC
}
\description{
Create a new sampleQC object
}
