% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{create_sge_object}
\alias{create_sge_object}
\title{Create a new SGE object}
\usage{
create_sge_object(file_libcount, file_allcount, file_valiant_meta)
}
\arguments{
\item{libcount}{QUANTS library-dependent count file, per sequence per count}

\item{allcount}{QUANTS library-independent count file, per sequence per count}

\item{valiant_meta}{VaLiAnT meta file}
}
\value{
An object of class SGE
}
\description{
Create a new SGE object
}
