% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{experimentQC}
\alias{experimentQC}
\title{A class representing a primary qc object}
\description{
A class representing a primary qc object
}
\section{Slots}{

\describe{
\item{\code{samples}}{a list of SGE objects}

\item{\code{coldata}}{a data frame of coldata for DESeq2}

\item{\code{ref_condition}}{the reference condition, like D4, others VS D4 in DESeq2}

\item{\code{library_counts_anno}}{a data frame of library counts of all the samples, annotated with consequences}

\item{\code{library_counts_pos_anno}}{a data frame of library counts of all the samples, annotated with consequences, sorted by position}

\item{\code{comparisons}}{a list of comparisons for degComps}

\item{\code{deseq_rlog}}{a data frame of deseq rlog counts of all the samples using library counts}

\item{\code{deseq_res}}{a list of deseq results of all the comparison against reference}
}}

