% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{format_count,SGE-method}
\alias{format_count,SGE-method}
\title{format library dependent and independent counts with extra info and remove duplicate/useless info}
\usage{
\S4method{format_count}{SGE}(object)
}
\arguments{
\item{object}{SGE object}
}
\value{
object
}
\description{
format library dependent and independent counts with extra info and remove duplicate/useless info
}
