% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{format_count}
\alias{format_count}
\title{format counts with extra info and remove duplicate/useless info}
\usage{
format_count(df_count)
}
\arguments{
\item{df_count}{input count dataframe}
}
\value{
a dataframe
}
\description{
format counts with extra info and remove duplicate/useless info
}
