% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expanded.R
\name{ggpairs_ext}
\alias{ggpairs_ext}
\title{ggpairs extended function}
\usage{
ggpairs_ext(data, mapping, pts = list(), smt = list(), ...)
}
\arguments{
\item{data}{Data}

\item{mapping}{Mapping}
}
\description{
the extension function of ggpair
}
