% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{import_sge_files}
\alias{import_sge_files}
\title{import files from the sample sheet
create objects based on the sample names}
\usage{
import_sge_files(
  dir_path,
  sample_sheet,
  file_libcount_hline = 3,
  file_allcount_hline = 3,
  file_valiant_meta_hline = 1,
  file_vep_anno_hline = 1,
  file_libcount_cols = vector(),
  file_allcount_cols = vector(),
  file_valiant_meta_cols = vector(),
  file_vep_anno_cols = vector()
)
}
\arguments{
\item{dir_path}{the directory path}

\item{sample_sheet}{the file name of the sample sheet which is in the directory}

\item{file_libcount_hline}{line number of header in library-dependent count file}

\item{file_allcount_hline}{line number of header in library-independent count file}

\item{file_valiant_meta_hline}{line number of header in VaLiAnT meta file}

\item{file_vep_anno_hline}{line number of header in vep annotation file}

\item{file_libcount_cols}{a vector of numbers of selected columns in library-dependent count file, default is none}

\item{file_allcount_cols}{a vector of numbers of selected columns in library-independent count file, default is none}

\item{file_valiant_meta_cols}{a vector of numbers of selected columns in VaLiAnT meta file, default is none}

\item{file_vep_anno_cols}{a vector of numbers of selected columns in vep annotation file, default is none}
}
\value{
a list of objects
}
\description{
import files from the sample sheet
create objects based on the sample names
}
