% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expanded.R
\name{merge_list_to_df}
\alias{merge_list_to_df}
\title{merge a list of vector values into a data frame
values may have different names in the list}
\usage{
merge_list_to_df(list_vals)
}
\arguments{
\item{objects}{a list of vector values}
}
\value{
a data frame
}
\description{
merge a list of vector values into a data frame
values may have different names in the list
}
