% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{primaryQC}
\alias{primaryQC}
\title{A class representing a primary qc object}
\description{
A class representing a primary qc object
}
\section{Slots}{

\describe{
\item{\code{samples}}{a list of SGE objects}

\item{\code{samples_ref}}{a list of SGE objects which are the references for screen QC}

\item{\code{samples_ref_filtered_seqs}}{a vector of sequences with counts > clustering cutoff for screen QC in the references}

\item{\code{counts}}{a list of sample counts}

\item{\code{counts_filtered}}{a data frame of filtered sample counts}

\item{\code{seq_clusters}}{1D k-means clusters, a data frame of sequences and cluster IDs}

\item{\code{stats}}{a data frame of samples and stats, eg. total no, filtered no.}
}}

