% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_outs.R
\name{qcout_sampleqc_length,sampleQC-method}
\alias{qcout_sampleqc_length,sampleQC-method}
\title{create output file of total reads stats}
\usage{
\S4method{qcout_sampleqc_length}{sampleQC}(object, len_bins = seq(0, 300, 50), outdir = NULL)
}
\arguments{
\item{object}{sampleQC object}

\item{len_bins}{the bins of length distribution}

\item{outdir}{the output directory}
}
\description{
create output file of total reads stats
}
