% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_outs.R
\name{qcout_sampleqc_library,sampleQC-method}
\alias{qcout_sampleqc_library,sampleQC-method}
\title{create output file of library reads stats}
\usage{
\S4method{qcout_sampleqc_library}{sampleQC}(object, outdir = NULL)
}
\arguments{
\item{object}{sampleQC object}

\item{outdir}{the output directory}
}
\description{
create output file of library reads stats
}
