% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_outs.R
\name{qcout_sampleqc_pos_cov,sampleQC-method}
\alias{qcout_sampleqc_pos_cov,sampleQC-method}
\title{create output file of lof percentages}
\usage{
\S4method{qcout_sampleqc_pos_cov}{sampleQC}(object, qctype = "screen", outdir = NULL)
}
\arguments{
\item{object}{sampleQC object}

\item{qctype}{screen or plasmid}

\item{outdir}{the output directory}
}
\description{
create output file of lof percentages
}
