% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_outs.R
\name{qcout_sampleqc_stats,sampleQC-method}
\alias{qcout_sampleqc_stats,sampleQC-method}
\title{create output file of bad seqs which fail filtering}
\usage{
\S4method{qcout_sampleqc_stats}{sampleQC}(object, outdir)
}
\arguments{
\item{object}{sampleQC object}

\item{outdir}{the output directory}
}
\description{
create output file of bad seqs which fail filtering
}
