% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_outs.R
\name{qcout_samqc_libcov,sampleQC-method}
\alias{qcout_samqc_libcov,sampleQC-method}
\title{create output file of library coverage}
\usage{
\S4method{qcout_samqc_libcov}{sampleQC}(object, outdir = NULL)
}
\arguments{
\item{object}{sampleQC object}

\item{outdir}{the output directory}
}
\description{
create output file of library coverage
}
