% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{qcplot_clusters,primaryQC-method}
\alias{qcplot_clusters,primaryQC-method}
\title{create the sequence counts and clusters plot}
\usage{
\S4method{qcplot_clusters}{primaryQC}(object, qctype, plotdir = "plotdir")
}
\arguments{
\item{object}{primaryQC object}

\item{qctype}{qc type for plot}

\item{plotdir}{the output plot directory}
}
\description{
create the sequence counts and clusters plot
}
