% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{qcplot_clusters,sampleQC-method}
\alias{qcplot_clusters,sampleQC-method}
\title{create the sequence counts and clusters plot}
\usage{
\S4method{qcplot_clusters}{sampleQC}(object, qctype = "screen", plotdir = NULL)
}
\arguments{
\item{object}{sampleQC object}

\item{qctype}{qc type for plot}

\item{plotdir}{the output plot directory}
}
\description{
create the sequence counts and clusters plot
}
