% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{qcplot_deseq_fc,sampleQC-method}
\alias{qcplot_deseq_fc,sampleQC-method}
\title{create fold change and consequence plot}
\usage{
\S4method{qcplot_deseq_fc}{sampleQC}(
  object,
  cons = c("synonymous", "lof", "missense"),
  pcut = 0.05,
  dcut = 0,
  ecut = 0,
  plotdir
)
}
\arguments{
\item{object}{sampleQC object}

\item{cons}{a vector of consequences showed in the figure}

\item{pcut}{the padj cutoff}

\item{dcut}{the depleted cutoff}

\item{ecut}{the enriched cutoff}

\item{plotdir}{the output plot directory}
}
\description{
create fold change and consequence plot
}
