% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{qcplot_dist_samples,experimentQC-method}
\alias{qcplot_dist_samples,experimentQC-method}
\title{create the heatmap of samples}
\usage{
\S4method{qcplot_dist_samples}{experimentQC}(object, plotdir = NULL)
}
\arguments{
\item{object}{experimentQC object}

\item{plotdir}{the output plot directory}
}
\description{
create the heatmap of samples
}
