% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{qcplot_dist_samples,sampleQC-method}
\alias{qcplot_dist_samples,sampleQC-method}
\title{create the heatmap of samples}
\usage{
\S4method{qcplot_dist_samples}{sampleQC}(object, plotdir)
}
\arguments{
\item{object}{sampleQC object}

\item{plotdir}{the output plot directory}
}
\description{
create the heatmap of samples
}
