% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{qcplot_expqc_deseq_fc,experimentQC-method}
\alias{qcplot_expqc_deseq_fc,experimentQC-method}
\title{create fold change and consequence plot}
\usage{
\S4method{qcplot_expqc_deseq_fc}{experimentQC}(
  object,
  cons = c("Synonymous_Variant", "LOF", "Missense_Variant"),
  pcut = 0.05,
  dcut = 0,
  ecut = 0,
  plotdir = NULL
)
}
\arguments{
\item{object}{experimentQC object}

\item{cons}{a vector of consequences showed in the figure}

\item{pcut}{the padj cutoff}

\item{dcut}{the depleted cutoff}

\item{ecut}{the enriched cutoff}

\item{plotdir}{the output plot directory}
}
\description{
create fold change and consequence plot
}
