% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{qcplot_expqc_sample_pca,experimentQC-method}
\alias{qcplot_expqc_sample_pca,experimentQC-method}
\title{create the pca of samples}
\usage{
\S4method{qcplot_expqc_sample_pca}{experimentQC}(object, ntop = 500, plotdir = NULL)
}
\arguments{
\item{object}{experimentQC object}

\item{ntop}{the number of top variances}

\item{plotdir}{the output plot directory}
}
\description{
create the pca of samples
}
