% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{qcplot_gini,sampleQC-method}
\alias{qcplot_gini,sampleQC-method}
\title{create the gini plot}
\usage{
\S4method{qcplot_gini}{sampleQC}(object, plotdir)
}
\arguments{
\item{object}{sampleQC object}

\item{plotdir}{the output plot directory}
}
\description{
create the gini plot
}
