% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{qcplot_pca_samples,sampleQC-method}
\alias{qcplot_pca_samples,sampleQC-method}
\title{create the pca of samples}
\usage{
\S4method{qcplot_pca_samples}{sampleQC}(object, ds_coldata, ntop = 500, plotdir)
}
\arguments{
\item{object}{sampleQC object}

\item{ds_coldata}{deseq coldata}

\item{ntop}{the number of top variances}

\item{plotdir}{the output plot directory}
}
\description{
create the pca of samples
}
