% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{qcplot_position,primaryQC-method}
\alias{qcplot_position,primaryQC-method}
\title{create the position plot}
\usage{
\S4method{qcplot_position}{primaryQC}(object, plotdir)
}
\arguments{
\item{object}{primaryQC object}

\item{plotdir}{the output plot directory}
}
\description{
create the position plot
}
