% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{qcplot_position,sampleQC-method}
\alias{qcplot_position,sampleQC-method}
\title{create the position plot}
\usage{
\S4method{qcplot_position}{sampleQC}(object, qctype = "screen", plotdir = NULL)
}
\arguments{
\item{object}{sampleQC object}

\item{qctype}{plot type, screen or plasmid}

\item{plotdir}{the output plot directory}
}
\description{
create the position plot
}
