% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{qcplot_position_anno,sampleQC-method}
\alias{qcplot_position_anno,sampleQC-method}
\title{create the position plot}
\usage{
\S4method{qcplot_position_anno}{sampleQC}(object, samples, type, major_cut = 0.005, plotdir)
}
\arguments{
\item{object}{sampleQC object}

\item{samples}{a vector of sample names}

\item{type}{plot type, lof or all}

\item{major_cut}{the qc cutoff of major variants}

\item{plotdir}{the output plot directory}
}
\description{
create the position plot
}
