% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{qcplot_samqc_total,sampleQC-method}
\alias{qcplot_samqc_total,sampleQC-method}
\title{create the stats plot}
\usage{
\S4method{qcplot_samqc_total}{sampleQC}(object, plotdir = NULL)
}
\arguments{
\item{object}{sampleQC object}

\item{plotdir}{the output plot directory}
}
\description{
create the stats plot
}
