% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{qcplot_stats,primaryQC-method}
\alias{qcplot_stats,primaryQC-method}
\title{create the stats plot}
\usage{
\S4method{qcplot_stats}{primaryQC}(object, plotdir = "plotdir")
}
\arguments{
\item{object}{primaryQC object}

\item{plotdir}{the output plot directory}
}
\description{
create the stats plot
}
