% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{read_count_file}
\alias{read_count_file}
\title{check file format and read it
file has header but beginning with #
reading extra comment info}
\usage{
read_count_file(file_path, file_type, hline)
}
\arguments{
\item{hline}{header line number}

\item{file}{file path}

\item{type}{library-dependent (lib) or library-independent (all)}
}
\value{
a dataframe
}
\description{
check file format and read it
file has header but beginning with #
reading extra comment info
}
