% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expanded.R
\name{round_any}
\alias{round_any}
\title{rounding function}
\usage{
round_any(x, accuracy, f = round)
}
\arguments{
\item{x}{X}

\item{accuracy}{ACC}
}
\value{
rounded x
}
\description{
Original plyr function for rounding
}
