% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experimentQC.R
\name{run_experiment_qc,experimentQC-method}
\alias{run_experiment_qc,experimentQC-method}
\title{run DESeq2 for the list of samples}
\usage{
\S4method{run_experiment_qc}{experimentQC}(object)
}
\arguments{
\item{object}{sampleQC object}
}
\value{
object
}
\description{
run DESeq2 for the list of samples
}
