% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleQC.R
\name{run_sample_qc,sampleQC-method}
\alias{run_sample_qc,sampleQC-method}
\title{run sample QC for the list of samples}
\usage{
\S4method{run_sample_qc}{sampleQC}(
  object,
  qc_type,
  library_count = 5,
  library_per = 0.25,
  cutoff_filtered = 1e+06,
  cutoff_mapping = 0.6,
  cutoff_library = 0.4,
  cutoff_cov = 100
)
}
\arguments{
\item{object}{sampleQC object}

\item{qc_type}{plasmid or screen}

\item{library_count}{count cutoff of library reads}

\item{library_per}{sample percentage cutoff of library reads}

\item{cutoff_filtered}{qc cutoff of the total filtered reads}

\item{cutoff_mapping}{qc cutoff of mapping percentage (ref + pam + library)}

\item{cutoff_library}{qc cutoff of library reads percentage}

\item{cutoff_cov}{qc cutoff of library coverage}
}
\value{
object
}
\description{
run sample QC for the list of samples
}
