% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleQC.R
\name{run_sample_qc_deseq2,sampleQC-method}
\alias{run_sample_qc_deseq2,sampleQC-method}
\title{run DESeq2 for the list of samples}
\usage{
\S4method{run_sample_qc_deseq2}{sampleQC}(object)
}
\arguments{
\item{object}{sampleQC object}
}
\value{
object
}
\description{
run DESeq2 for the list of samples
}
