% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{sampleQC}
\alias{sampleQC}
\title{A class representing a primary qc object}
\description{
A class representing a primary qc object
}
\section{Slots}{

\describe{
\item{\code{samples}}{a list of SGE objects}

\item{\code{samples_ref}}{a list of SGE objects which are the references for screen QC}

\item{\code{counts}}{a list of sample counts}

\item{\code{lengths}}{a list of sequence lengths}

\item{\code{seq_clusters}}{a list of dataframes of sequences and cluster IDs}

\item{\code{filtered_seqs}}{a vector of sequences with counts > clustering cutoff for screen QC in the references}

\item{\code{filtered_counts}}{a data frame of filtered counts of all the samples}

\item{\code{effective_counts}}{a data frame of effective counts of all the samples}

\item{\code{effective_counts_pos}}{a data frame of effective counts of all the samples sorted by position}

\item{\code{effective_counts_anno}}{a data frame of effective counts of all the samples, annotated with consequences}

\item{\code{effective_counts_pos_anno}}{a data frame of effective counts of all the samples, annotated with consequences, sorted by position}

\item{\code{stats}}{a data frame of samples and stats, eg. total no, filtered no.}

\item{\code{filtered_samples}}{a vector of filtered sample names}

\item{\code{bad_seqs}}{a list of filter-out sequences}

\item{\code{deseq_rlog}}{a data frame of deseq rlog counts of all the samples using effective counts}

\item{\code{deseq_res}}{a list of deseq results of all the comparison against reference}
}}

