% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expanded.R
\name{select_colorblind}
\alias{select_colorblind}
\title{color blind friendly}
\usage{
select_colorblind(col_id)
}
\arguments{
\item{col_id}{a character to select colors}
}
\value{
a vector of colors
}
\description{
color blind friendly
}
