% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expanded.R
\name{select_objects}
\alias{select_objects}
\title{fetch objects from list by the names or indexes}
\usage{
select_objects(objects, tags)
}
\arguments{
\item{objects}{a list of objects}

\item{tags}{a vector of names or indexes}
}
\value{
a list of objects
}
\description{
fetch objects from list by the names or indexes
}
