% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{SGE}
\alias{SGE}
\title{A class representing a SGE object}
\description{
A class representing a SGE object
}
\section{Slots}{

\describe{
\item{\code{sample}}{name}

\item{\code{library}}{type}

\item{\code{library}}{name}

\item{\code{QUANTS}}{library-dependent count file, per sequence per count}

\item{\code{QUANTS}}{library-independent count file, per sequence per count}

\item{\code{VaLiAnT}}{meta file}
}}

