% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{sge_stats,SGE-method}
\alias{sge_stats,SGE-method}
\title{format library dependent and independent counts with extra info and remove duplicate/useless info}
\usage{
\S4method{sge_stats}{SGE}(object, lowcut = 10)
}
\arguments{
\item{object}{SGE object}

\item{lowcut}{cutoff which determines the oligo count is low, user's definition}
}
\value{
object
}
\description{
format library dependent and independent counts with extra info and remove duplicate/useless info
}
