% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{show,sampleQC-method}
\alias{show,sampleQC-method}
\title{show basic info of the object}
\usage{
\S4method{show}{sampleQC}(object)
}
\arguments{
\item{object}{primary object}
}
\description{
show basic info of the object
}
