% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{show_stats,SGE-method}
\alias{show_stats,SGE-method}
\title{show basic stats of the object}
\usage{
\S4method{show_stats}{SGE}(object)
}
\arguments{
\item{object}{SGE object}
}
\description{
show basic stats of the object
}
