% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display.R
\name{show_stats_qc,SGE-method}
\alias{show_stats_qc,SGE-method}
\title{show qc stats of the object}
\usage{
\S4method{show_stats_qc}{SGE}(object)
}
\arguments{
\item{object}{SGE object}
}
\description{
show qc stats of the object
}
