% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expanded.R
\name{t_col}
\alias{t_col}
\title{transparent color function}
\usage{
t_col(col, rate)
}
\arguments{
\item{col}{color name}

\item{rate}{alpha rate}
}
\value{
transparent color
}
\description{
transparent color function
}
