% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expanded.R
\name{trim_adaptor}
\alias{trim_adaptor}
\title{trim adaptor sequences}
\usage{
trim_adaptor(seq, adapt5, adapt3)
}
\arguments{
\item{seq}{sequence}

\item{adapt5}{5 prime adaptor sequence}

\item{adapt3}{3 prime adaptor sequence}
}
\value{
string
}
\description{
trim adaptor sequences
}
