% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{qcplot_position_anno,sampleQC-method}
\alias{qcplot_position_anno,sampleQC-method}
\title{create the position plot}
\usage{
\S4method{qcplot_position_anno}{sampleQC}(object, samples = NULL, type = "lof", plotdir = NULL)
}
\arguments{
\item{object}{sampleQC object}

\item{samples}{a vector of sample names}

\item{type}{plot type, lof or all}

\item{plotdir}{the output plot directory}
}
\description{
create the position plot
}
