% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc_plots.R
\name{qcplot_readlens,sampleQC-method}
\alias{qcplot_readlens,sampleQC-method}
\title{create the read length plot}
\usage{
\S4method{qcplot_readlens}{sampleQC}(object, len_bins = seq(0, 300, 50), plotdir = NULL)
}
\arguments{
\item{object}{sampleQC object}

\item{len_bins}{the bins of length distribution}

\item{plotdir}{the output plot directory}
}
\description{
create the read length plot
}
