% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{read_sge_file}
\alias{read_sge_file}
\title{check file format and read it
file has header but without # at the beginning
ignore all the comments}
\usage{
read_sge_file(file_path)
}
\arguments{
\item{file}{file path}
}
\value{
a dataframe
}
\description{
check file format and read it
file has header but without # at the beginning
ignore all the comments
}
