% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ingest.R
\name{read_sge_file}
\alias{read_sge_file}
\title{check file format and read it
file has header but beginning with #}
\usage{
read_sge_file(file_path, hline = 0, colnums = vector())
}
\arguments{
\item{file_path}{the file path}

\item{hline}{header line number, default is 0}

\item{colnums}{a vector of selected colummns, default is none}
}
\value{
a dataframe
}
\description{
check file format and read it
file has header but beginning with #
}
