% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/primaryQC.R
\name{run_primary_qc,primaryQC-method}
\alias{run_primary_qc,primaryQC-method}
\title{run primary QC for the list of samples}
\usage{
\S4method{run_primary_qc}{primaryQC}(object, qc_type, cluster_cutoff)
}
\arguments{
\item{object}{primaryQC object}

\item{qc_type}{plasmid or screen}
}
\value{
object
}
\description{
run primary QC for the list of samples
}
