% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{sampleQC}
\alias{sampleQC}
\title{A class representing a primary qc object}
\description{
A class representing a primary qc object
}
\section{Slots}{

\describe{
\item{\code{cutoffs}}{a data frame of cutoffs using in sample QC}

\item{\code{samples}}{a list of SGE objects}

\item{\code{samples_ref}}{a list of SGE objects which are the references for screen QC}

\item{\code{counts}}{a list of sample counts}

\item{\code{lengths}}{a list of sequence lengths}

\item{\code{seq_clusters}}{a list of dataframes of sequences and cluster IDs}

\item{\code{accepted_seqs}}{a list of sequences with counts > clustering cutoff for screen QC in the references}

\item{\code{accepted_counts}}{a list of filtered counts of all the samples}

\item{\code{library_counts}}{a list of library counts of all the samples}

\item{\code{unmapped_counts}}{a list of unmapped counts against meta library sequences of all the samples}

\item{\code{library_counts_chr}}{a list of chromosomes for library counts}

\item{\code{library_counts_pos}}{a list of library counts of all the samples sorted by position in meta}

\item{\code{library_counts_anno}}{a data frame of library counts of all the samples, annotated with consequences}

\item{\code{library_counts_pos_anno}}{a data frame of library counts of all the samples, annotated with consequences, sorted by position}

\item{\code{stats}}{a data frame of samples and stats, eg. total no, filtered no.}

\item{\code{bad_seqs_bycluster}}{a list of filter-out sequences by cluster}

\item{\code{bad_seqs_bydepth}}{a list of filter-out sequences by depth}

\item{\code{bad_seqs_bylib}}{a list of filter-out sequences by library mapping}

\item{\code{filtered_samples}}{a vector of filtered sample names}
}}

