% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{SGE}
\alias{SGE}
\title{A class representing a SGE object}
\description{
A class representing a SGE object
}
\section{Slots}{

\describe{
\item{\code{samples}}{sample names}

\item{\code{targetons}}{targeton ids}

\item{\code{libtype}}{library type}

\item{\code{libname}}{library name}

\item{\code{libcounts}}{QUANTS library-dependent count file, per sequence per count}

\item{\code{allcounts}}{QUANTS library-independent count file, per sequence per count}

\item{\code{valiant_meta}}{VaLiAnT meta file}
}}

