% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{SGE}
\alias{SGE}
\title{A class representing a SGE object}
\description{
A class representing a SGE object
}
\section{Slots}{

\describe{
\item{\code{samples}}{sample names}

\item{\code{targetons}}{targeton ids}

\item{\code{libname}}{library name}

\item{\code{libtype}}{library type}

\item{\code{adapt5}}{adaptor sequence at 5 prime end}

\item{\code{adapt3}}{adaptor sequence at 3 prime end}

\item{\code{refseq}}{reference sequence}

\item{\code{pamseq}}{sequence with pam variants}

\item{\code{libcounts}}{QUANTS library-dependent count file, per sequence per count}

\item{\code{allcounts}}{QUANTS library-independent count file, per sequence per count}

\item{\code{valiant_meta}}{VaLiAnT meta file}

\item{\code{libstats}}{summaries of library dependent counts}

\item{\code{allstats}}{summaries of library independent counts}
}}

