% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{SGE}
\alias{SGE}
\title{A class representing a SGE object}
\description{
A class representing a SGE object
}
\section{Slots}{

\describe{
\item{\code{sample}}{the sample name}

\item{\code{libname}}{library name}

\item{\code{libtype}}{library type}

\item{\code{adapt5}}{adaptor sequence at 5 prime end}

\item{\code{adapt3}}{adaptor sequence at 3 prime end}

\item{\code{refseq}}{reference sequence}

\item{\code{pamseq}}{sequence with pam variants}

\item{\code{libcounts}}{QUANTS library-dependent counts, per sequence per count}

\item{\code{allcounts}}{QUANTS library-independent counts, per sequence per count}

\item{\code{valiant_meta}}{VaLiAnT meta file}

\item{\code{vep_anno}}{vep consequence annotation file}

\item{\code{meta_mseqs}}{non-redundant mseq in VaLiAnT meta file}

\item{\code{missing_meta_seqs}}{missing sequenced in library compared to VaLiAnT meta file}

\item{\code{libstats}}{summaries of library dependent counts}

\item{\code{allstats}}{summaries of library independent counts}

\item{\code{libstats_qc}}{qc stats of library dependent counts}

\item{\code{allstats_qc}}{qc stats of library independent counts}
}}

